+++ src/lib/rxfm/components/html.ts
@@ -3,7 +3,7 @@ import { componentCreator, ComponentCreator, componentFunction } from "./compone
 /**
  * A function to return a component creator for an HTML element with the given tag name.
  */
-function htmlComponentCreator<K extends keyof HTMLElementTagNameMap>(
+export function htmlComponentCreator<K extends keyof HTMLElementTagNameMap>(
   tagName: K,
 ): ComponentCreator<HTMLElementTagNameMap[K]> {
   return componentCreator(
@@ -14,14 +14,12 @@ function htmlComponentCreator<K extends keyof HTMLElementTagNameMap>(
 export const A = htmlComponentCreator('a');
 export const Abbr = htmlComponentCreator('abbr');
 export const Address = htmlComponentCreator('address');
-export const Applet = htmlComponentCreator('applet');
 export const Area = htmlComponentCreator('area');
 export const Article = htmlComponentCreator('article');
 export const Aside = htmlComponentCreator('aside');
 export const Audio = htmlComponentCreator('audio');
 export const B = htmlComponentCreator('b');
 export const Base = htmlComponentCreator('base');
-export const Basefont = htmlComponentCreator('basefont');
 export const Bdi = htmlComponentCreator('bdi');
 export const Bdo = htmlComponentCreator('bdo');
 export const Blockquote = htmlComponentCreator('blockquote');