+++ src/lib/rxfm/rxfm-jsx/index.ts
@@ -1,0 +1,3 @@
+export * from './rxfm-jsx';
+import RxFM from './rxfm-jsx';
+export default RxFM;