+++ webpack.app.js
@@ -5,13 +5,13 @@ module.exports = [
   {
     mode: "production",
     devtool: "source-map",
-    entry: "./src/app/index.ts",
+    entry: "./src/app/index.tsx",
     output: {
       filename: 'app.bundle.js',
       path: path.resolve(__dirname, 'dist'),
     },
     resolve: {
-      extensions: [".ts", ".js", ".json"],
+      extensions: [".ts", ".tsx", ".js", ".jsx", ".json"],
       alias: {
         "rxfm": path.resolve(__dirname, "./src/lib/rxfm/index")
       },
@@ -19,7 +19,7 @@ module.exports = [
     module: {
       rules: [
         {
-          test: /\.ts$/,
+          test: /\.ts(x)?$/,
           loader: "ts-loader",
           exclude: '/node_modules/'
         },